update part 3 & 4 of R2

+++ R2/src/main/java/calculator/Calculator.java
@@ -6,11 +6,11 @@ public class Calculator {
         return a + b;
     }
 
-    public float reverse(int a) {
+    public float reverse(float a) {
         return (float) 1 / a;
     }
 
-    public float sqrt(int a) {
+    public float sqrt(float a) {
         return (float) Math.sqrt(a);
     }
 }
